Move isNodeRecordWithChildren to nodeUtils; add test
