+++ src/nodeUtils.ts
@@ -1,0 +1,10 @@
+interface NodeRecordWithChildren extends NodeRecord {
+    children: NodeData[];
+}
+
+export const isNodeRecordWithChildren = (
+    data: NodeData,
+): data is NodeRecordWithChildren =>
+    typeof data === "object" &&
+    "children" in data &&
+    data["children"] instanceof Array;
